TST: remove test of DOC and rename streamer to sinks (these handle our databroker inserts)
